refactor(ProCivOccurrenceStatuses): use TINYINT for codes column

+++ database/migrations/2019_04_17_203847_create_prociv_occurrence_statuses_table.php
@@ -18,7 +18,7 @@ class CreateProcivOccurrenceStatusesTable extends Migration
         Schema::create('prociv_occurrence_statuses', function (Blueprint $table) {
             $table->tinyIncrements('id');
 
-            $table->unsignedSmallInteger('code')->unique();
+            $table->unsignedTinyInteger('code')->unique();
             $table->string('name')->unique();
 
             $table->timestamps();
